+++ src/UKP/UKPmanager.hpp
@@ -15,18 +15,16 @@ private:
 #if SEVEN_SEGMENT
     _7segment seg;
 #endif
-// if profiles are enabled track when they change
-#if PROFILES
+#if PROFILES // Profile behavior.
     int currentProfile;
     int lastProfileState;
     void profileChanged();
     inline void changeProfile();
 #endif
 #ifdef KILL_SWITCH
-    boolean killSwitch();
+    bool killSwitch();
 #endif
     void manageButtonMacros(int &buttonId);
-    int pinTriggered; // save the button id of the last pressed button
 public:
     void begin();
     void runtime();