+++ src/settings.h
@@ -43,9 +43,7 @@
 #endif
 
 //define macros
-
 #define ARR_SIZE(X) sizeof(X)/sizeof(X[0])
-#define IF_TRUE(X) X ? 1 : 0
 #define RETRIEVE_PROFILE(POS) (char *)pgm_read_word(&(defaultMacros[POS]))
 #define MACRO const char PROGMEM
 #define MACRO_ARRAY const char *const PROGMEM